requirements.txt
aiodns==3.0.0
aiohttp==3.9.0b0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
crispy-tailwind==1.0.3
cycler==0.12.1
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-address==0.2.8
django-crispy-forms==2.1
django-filter==23.5
django-jazzmin==3.0.0
django-light==0.1.0.post3
django-multiselectfield==0.1.12
django-phonenumber-field==7.3.0
django-six==1.0.5
django-smtp-ssl==1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
fonttools==4.48.1
frozenlist==1.4.1
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
googlemaps==4.10.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpld3==0.5.10
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openai==1.12.0
packaging==23.2
pandas==2.2.1
phonenumbers==8.13.30
pillow==10.2.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stripe==8.2.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
zipp==3.17.0
